+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/requests/chat/modify/UnpinChatMessage.kt
@@ -3,14 +3,18 @@ package dev.inmo.tgbotapi.requests.chat.modify
 import dev.inmo.tgbotapi.CommonAbstracts.types.ChatRequest
 import dev.inmo.tgbotapi.requests.abstracts.SimpleRequest
 import dev.inmo.tgbotapi.types.ChatIdentifier
+import dev.inmo.tgbotapi.types.MessageIdentifier
 import dev.inmo.tgbotapi.types.chatIdField
+import dev.inmo.tgbotapi.types.messageIdField
 import kotlinx.serialization.*
 import kotlinx.serialization.builtins.serializer
 
 @Serializable
 data class UnpinChatMessage(
     @SerialName(chatIdField)
-    override val chatId: ChatIdentifier
+    override val chatId: ChatIdentifier,
+    @SerialName(messageIdField)
+    val messageId: MessageIdentifier? = null
 ): ChatRequest, SimpleRequest<Boolean> {
     override fun method(): String = "unpinChatMessage"
     override val resultDeserializer: DeserializationStrategy<Boolean>